+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## Unreleased
+
+### Changed
+
+- Moved `offline: false` parameter to `print_to_pdf` options so it can be switched on and off.
+- Targets now navigate to `about:blank` after PDF prints.
+
 ## [0.1.0] - 2020-04-02
 
 ### Added